add features

+++ engine/source/runtime/core/math/vector2i.cpp
@@ -10,7 +10,7 @@ namespace lain
     const Vector2i Vector2i::NEGATIVE_UNIT_Y(0, -1);
     const Vector2i Vector2i::UNIT_SCALE(1, 1);
     Vector2i::operator Vector2() const {
-        return Vector2((int32_t)x, (int32_t)y);
+        return Vector2(static_cast<float>(x), static_cast<float>(y));
     }
